--- cloudshell/networking/core/apply_connectivity/connectivity_action_result.py
@@ -1,9 +1,0 @@
-class ConnectivityActionResult(object):
-
-    def __init__(self):
-        self.actionId = ''
-        self.type = ''
-        self.infoMessage = ''
-        self.errorMessage = ''
-        self.success = True
-        self.updatedInterface = ''
